refactor: 쿠폰 이름 객체 분리
