+++ src/main/java/coupon/coupon/domain/CouponName.java
@@ -1,0 +1,33 @@
+package coupon.coupon.domain;
+
+import java.util.Objects;
+import coupon.coupon.CouponException;
+
+public class CouponName {
+
+    private static final int MAX_LENGTH = 30;
+
+    private final String name;
+
+    public CouponName(String name) {
+        validate(name);
+        this.name = name;
+    }
+
+    private void validate(String name) {
+        validateNull(name);
+        validateNameLength(name);
+    }
+
+    private void validateNull(String name) {
+        if (Objects.isNull(name)) {
+            throw new CouponException("쿠폰 이름이 누락되었습니다.");
+        }
+    }
+
+    private void validateNameLength(String name) {
+        if (name.isEmpty() || name.length() > MAX_LENGTH) {
+            throw new CouponException("쿠폰은 30자 이하의 이름을 설정해주세요.");
+        }
+    }
+}
